requirements_before_pytorch_cpu.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
arrow==1.4.0
asgiref==3.10.0
attrs==25.4.0
audioread==3.1.0
autobahn==25.10.2
Automat==25.4.16
baidu-aip==4.16.13
bcrypt==5.0.0
betterproto==2.0.0b6
binaryornot==0.4.4
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cmake==4.1.2
coloredlogs==15.0.1
constantly==23.10.4
cookiecutter==2.6.0
coverage==7.11.1
crcmod==1.7
cryptography==41.0.0
daphne==4.2.1
dashscope==1.14.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docarray==0.41.0
ecdsa==0.19.1
editdistance==0.8.1
editdistance-s==1.0.0
email-validator==2.3.0
fastapi==0.110.0
filelock==3.20.0
flake8==6.0.0
flatbuffers==25.9.23
flet==0.27.6
flet-cli==0.27.6
flet-desktop-light==0.27.6
flet-video==0.1.1
frozenlist==1.8.0
fsspec==2025.10.0
funasr==1.2.0
greenlet==3.2.4
grpclib==0.4.8
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
iniconfig==2.3.0
jaconv==0.3
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
jmespath==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kaldiio==2.18.1
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-huggingface==1.0.1
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lazy_loader==0.4
librosa==0.10.0
lit==18.1.8
llvmlite==0.42.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
mini-racer==0.12.4
ml_dtypes==0.5.3
modelscope==1.31.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
networkx==3.5
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.19.0
onnxconverter-common==1.16.0
onnxruntime==1.23.0
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
oss2==2.18.0
packaging==24.2
pandas==2.1.4
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
plyer==2.1.0
pooch==1.8.2
priority==2.0.0
prometheus-client==0.17.0
propcache==0.4.1
protobuf==6.33.0
psutil==5.9.6
py-mini-racer==0.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.11
pycodestyle==2.10.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydub==0.25.1
PyExecJS==1.5.1
pyflakes==3.0.1
Pygments==2.19.2
PyMySQL==1.1.0
pynndescent==0.5.13
pyOpenSSL==25.3.0
pypng==0.20220715.0
pystray==0.19.5
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
python-xlib==0.33
pytokens==0.3.0
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
redis-py-cluster==2.1.3
regex==2025.11.3
repath==0.9.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.10.0
screeninfo==0.8.1
sentence-transformers==3.3.0
sentencepiece==0.1.99
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
snownlp==0.12.3
soundfile==0.12.0
soxr==1.0.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.0
srt==3.5.0
starlette==0.36.3
streamget==4.0.8
sympy==1.13.1
tenacity==8.5.0
tensorboardX==2.6.4
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch-complex==0.4.4
tqdm==4.65.0
transformers==4.57.1
triton==2.2.0
Twisted==25.5.0
txaio==25.9.2
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.0
urllib3==2.0.7
uv==0.9.8
uvicorn==0.23.0
uvloop==0.22.1
uWSGI==2.0.31
watchdog==4.0.2
watchfiles==1.1.1
websocket-client==1.6.1
websockets==11.0.2
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zope.interface==8.0.1
zstandard==0.25.0
